meeting-detail controller 삭제 및 레파지토리 이동
