--- src/meeting-detail/meeting-detail.service.ts
@@ -1,19 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { MeetingDetailRepository } from './repositories/meeting-detail.repository';
-
-@Injectable()
-export class MeetingDetailService {
-  constructor(
-    @InjectRepository(MeetingDetailRepository)
-    private meetingDetailRepository: MeetingDetailRepository,
-  ) {}
-
-  // 멤버 기록 보기
-  async getMemberRecord(meeting_id: number, member_id: number) {
-    return {
-      status: 'SUCCESS',
-      code: 200,
-    };
-  }
-}
