+++ src/meeting/meeting.module.ts
@@ -3,6 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { CategoryRepository } from 'src/category/repositories/category.repository';
 import { MeetingController } from './meeting.controller';
 import { MeetingService } from './meeting.service';
+import { MeetingDetailRepository } from './repositories/meeting-detail.repository';
 import { MeetingUserRepository } from './repositories/meeting-user.repository';
 import { MeetingRepository } from './repositories/meeting.repository';
 
@@ -12,6 +13,7 @@ import { MeetingRepository } from './repositories/meeting.repository';
       MeetingRepository,
       MeetingUserRepository,
       CategoryRepository,
+      MeetingDetailRepository,
     ]),
   ],
   controllers: [MeetingController],
