+++ src/meeting/repositories/meeting-detail.repository.ts
